Sort the cache in the Loader::getRequiredFile method.
The idea is that regardless how many paths we have to check, only a few
are really used (in my system only one :P). So, in the for loop that checks
the standard search path, the cache of urls gets sorted, putting the used path
as the first element of the cache. This way, the most used paths will end up
in the beginning of the cache, and therefore, we'll return sooner next time.