Merged the ClassDeclaration and the ModuleDeclaration classes.
In the DUChain's point of view, Ruby classes and modules are almost the same
structure. The only difference is that classes may have a base class. By
merging both classes and adding a new "isModule" attribute to the data, we
are able to finally differentiate between both structures (previously there
were some crazy bugs for this topic).