Partial re-write of the lexer.
This commit contains a partial re-write of the lexer that can be found in
the parser/parser.y file. Tests have been updated. This time the lexer is
structured in a giant switch statement, making it faster and smaller. Moreover,
this commit comes handy for future optimizations on the parser.

Note also that the support for heredocs has been removed. This is because the
current support is broken, and I plan to provide a proper implementation in
the near future.