Greatly reduced the push/pop operations around positions.
Since this plugin doesn't need to know the position of *every* single
node, we now just store the strictly necessary ones. This way the number
of push/pop operations on the positions stack gets greatly reduced. Moreover,
a lot of code that fixed things around positions is no longer needed.