Adapt to kdevplatform's PathFilter API changes
https://commits.kde.org/kdevelop/9d55d42b3665f1bc1fb7cba6350351ce8250f00c
removed duplicate calls to PathFilter::clearFilter() and made this
member function private. The clearFilter() call in DataProvider::reset()
is duplicate for the same reason as in KDevelop => remove it too. The
following compilation error (GCC) is fixed by the removal:
  ‘void KDevelop::PathFilter<Item, Parent>::clearFilter() [with ...]’ is private within this context

https://commits.kde.org/kdevelop/b37a289d22e7717795307b038d3fedaabe4301ac
changed the signature of PathFilter::setItems() and renamed it to
updateItems() to improve performance => adapt and refactor
DataProvider::reset() in the same way as analogous code in KDevelop's
OpenFilesDataProvider::reset() was modified. This change fixes the
following compilation error (GCC):
  ‘setItems’ was not declared in this scope